refactor models

+++ src/main/java/com/example/demo/repositories/CreditDetailRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repositories;
 import com.example.demo.models.CreditDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CreditDetailRepository extends JpaRepository<CreditDetail, Integer> {}
+public interface CreditDetailRepository extends JpaRepository<CreditDetail, Long> {}
