+++ src/main/java/com/example/demo/repositories/RegisterReceiptRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repositories;
 import com.example.demo.models.RegisterReceipt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegisterReceiptRepository extends JpaRepository<RegisterReceipt, Integer> {}
+public interface RegisterReceiptRepository extends JpaRepository<RegisterReceipt, Long> {}
