+++ src/main/java/com/example/demo/repositories/SubjectScheduleRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.repositories;
 import com.example.demo.models.SubjectSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubjectScheduleRepository extends JpaRepository<SubjectSchedule, Integer> {}
+public interface SubjectScheduleRepository extends JpaRepository<SubjectSchedule, Long> {}
